Split segment_reduction_ops_gpu.cu.cc compilation
- Refactors the file into a header and several separate source files
  to reduce critical path compile time. (This is done pre-emptively
  because future changes that are expected to increase the compile
  time of this code).
- No functional change.
